Began Multi-threaded support
Implemented ForkJoinPool execution to test 2 (low precision). Created a
BasicTest class to replace test1. A BasicTest object handles only one
pixel. The test2 method of the Tester class is now used to create
BasicTest objects and submitting them to the pool. Instead of returning
a resultArray, the BasicTest object will edit the resultsArray in the
Launcher class when the test finishes. The main thread loops in test2
until all tasks in the pool have been completed, updating the progress
bar accordingly.