Added custom Gradients and color cycling
Created gradient point and gradient classes that are used to defined the
colors of the graph. Added a color cycle class that creates a new thread
and increments an offset that varies the color of each point on the
graph, then repaints the graph (does not recompute the mandelbrot set so
is only dependent on graph size).

Close #17