Updated cursor
changed cursor to be the hand cursor over the graph to make it more
intuitive to click on the graph. Also added logic to change the cursor
to the "wait cursor" while processing. The tester object will call
display.waitCursor(true); to change the cursor at the beginning and the
end of the test2 or test4 method.